+++ lib/data/Database.dart
@@ -1,8 +1,10 @@
 import 'package:moor/moor.dart';
 
+import 'dao/GroupDao.dart';
 import 'dao/NicknameDao.dart';
 import 'dao/SettingsDao.dart';
 import 'dao/ViveBaseStationDao.dart';
+import 'tables/GroupTable.dart';
 import 'tables/LastSeenDevicesTable.dart';
 import 'tables/NicknameTable.dart';
 import 'tables/SimpleSettingsTable.dart';
@@ -29,23 +31,32 @@ class NicknamesLastSeenJoin {
   LastSeenDevices,
   SimpleSettings,
   ViveBaseStationIds,
+  Groups,
+  GroupEntries,
 ], daos: [
   NicknameDao,
   SettingsDao,
-  ViveBaseStationDao
+  ViveBaseStationDao,
+  GroupDao,
 ])
 class LighthouseDatabase extends _$LighthouseDatabase {
   LighthouseDatabase(QueryExecutor e) : super(e);
 
   @override
-  int get schemaVersion => 2;
+  int get schemaVersion => 3;
 
   @override
   MigrationStrategy get migration => MigrationStrategy(onCreate: (Migrator m) {
         return m.createAll();
       }, onUpgrade: (Migrator m, int from, int to) async {
-        if (from == 1 && to == 2) {
+        if (from == 1 && (to >= 2 && to <= 3)) {
           await m.renameColumn(simpleSettings, 'id', simpleSettings.settingsId);
         }
+        if ((from >= 1 && from <= 2) && (to == 3)) {
+          await m.createTable(groups);
+          await m.createTable(groupEntries);
+        }
+      }, beforeOpen: (details) async {
+        await this.customStatement('PRAGMA foreign_keys = ON');
       });
 }