+++ lib/pages/settings/SettingsNicknamesPage.dart
@@ -79,9 +79,9 @@ class _NicknamesPageState extends State<_SettingsNicknamesPageContent> {
           }
         }
 
-        final Color scaffoldColor = selected.isNotEmpty
+        final Color? scaffoldColor = selected.isNotEmpty
             ? Theme.of(context).selectedRowColor
-            : Theme.of(context).primaryColor;
+            : null;
         final List<Widget> actions = selected.isEmpty
             ? const []
             : [